FEAT: update Retro components and styles for improved UI and functionality
