+++ client/src/pages/retro/RetroBoard.jsx
@@ -1,127 +1,73 @@
-// import Grid from '@mui/material/Grid2';
-// import { Typography, Container, Divider } from '@mui/material';
-
-// import './Retro.css';
-// import RetroMood from './RetroMood';
-// import RetroReview from './RetroReview';
-// import { useWebSocket } from '../../hooks/useWebSocket';
-// import AuthProvider from '../../middleware/AuthProvider';
-
-
-// export default function Retro() {
-
-//     const { token } = AuthProvider();
-//     const { retroData, updateEmoji, addComment } = useWebSocket();
-
-//     if (!token) {
-//         return <div>Please log in to access the retro board.</div>;
-//     }
-
-//     return (
-//         <Container maxWidth="xl">
-//             <div className="retro-header">
-//                 <div className="retro-header-title">
-//                     <Typography variant="h4" gutterBottom>
-//                         Retro-Board
-//                     </Typography>
-//                     <Typography variant="subtitle1" gutterBottom>
-//                         How do you feel about the sprint?
-//                     </Typography>
-//                 </div>
-//                 <div>
-//                     <RetroMood emojis={retroData?.emojis || null} updateEmoji={updateEmoji} />
-//                 </div>
-//             </div>
-
-//             <Divider />
-
-//             <Grid container spacing={1} mt={1}>
-//                 <RetroReview
-//                     title="Start Doing"
-//                     data={retroData?.comments?.startDoing || []}
-//                     addComment={(text) => addComment('startDoing', text)}
-//                 />
-//                 <RetroReview
-//                     title="Stop Doing"
-//                     data={retroData?.comments?.stopDoing || []}
-//                     addComment={(text) => addComment('stopDoing', text)}
-//                 />
-//                 <RetroReview
-//                     title="Continue Doing"
-//                     data={retroData?.comments?.continueDoing || []}
-//                     addComment={(text) => addComment('continueDoing', text)}
-//                 />
-//                 <RetroReview
-//                     title="Appreciation"
-//                     data={retroData?.comments?.appreciation || []}
-//                     addComment={(text) => addComment('appreciation', text)}
-//                 />
-//             </Grid>
-//         </Container>
-//     );
-// };
-
-
-
-
-
-
 import Grid from '@mui/material/Grid2';
-import { Container, Typography, Divider } from '@mui/material';
+import { Typography, Container, Divider, Card } from '@mui/material';
 
+import './Retro.css';
 import RetroMood from './RetroMood';
-import RetroColumn from './RetroColumn';
-import useRetroSocket from '../../hooks/useRetroSocket';
+import RetroReview from './RetroReview';
+import { useAuth } from '../../hooks/useAuth';
+import { useRetroSocket } from '../../hooks/useRetroSocket';
 
 
+// RetroBoard page component
 export default function RetroBoard() {
-    const retroId = '1234567890';
-    const { boardData, updateEmoji, addComment } = useRetroSocket(retroId);
+
+    const retroId = "0987654321";
+    const { isAuthenticated } = useAuth();
+    const { retroData, updateMood, addReview, updateReview } = useRetroSocket(retroId);
+
+    if (!isAuthenticated) {
+        return null;
+    }
 
     return (
-        <Container maxWidth="xl" sx={{ py: 3 }}>
-            {/* Header Section */}
-            <Typography variant="h3" gutterBottom>
-                Sprint Retrospective
-            </Typography>
-            <RetroMood
-                emojis={boardData?.emojis}
-                onSelect={updateEmoji}
-            />
-            <Divider sx={{ my: 3 }} />
+        <Container maxWidth="xl">
+            <div className="retro-header">
+                <div className="retro-header-title">
+                    <Typography variant="h4" gutterBottom>
+                        Retro-Board
+                    </Typography>
+                    <Typography variant="subtitle1" gutterBottom>
+                        How do you feel about the sprint?
+                    </Typography>
+                </div>
+                <div>
+                    <RetroMood moods={retroData?.moods ?? []} updateMood={updateMood} />
+                </div>
+            </div>
 
-            {/* Retro Columns */}
-            <Grid container spacing={3}>
-                <Grid item xs={12} md={3}>
-                    <RetroColumn
+            <Divider />
+
+            <Card raised elevation={1} sx={{ mt: 1, p: 1 }}>
+                <Grid container spacing={1}>
+                    <RetroReview
                         title="Start Doing"
-                        items={boardData?.comments?.startDoing || []}
-                        onAdd={(text) => addComment('startDoing', text)}
+                        data={retroData?.reviews?.startDoing || []}
+                        addReview={(text) => addReview('startDoing', text)}
+                        updateReview={(text, index) => updateReview('startDoing', text, index)}
                     />
-                </Grid>
-                <Grid item xs={12} md={3}>
-                    <RetroColumn
+
+                    <RetroReview
                         title="Stop Doing"
-                        items={boardData?.comments?.stopDoing || []}
-                        onAdd={(text) => addComment('stopDoing', text)}
+                        data={retroData?.reviews?.stopDoing || []}
+                        addReview={(text) => addReview('stopDoing', text)}
+                        updateReview={(text, index) => updateReview('stopDoing', text, index)}
                     />
-                </Grid>
-                <Grid item xs={12} md={3}>
-                    <RetroColumn
+
+                    <RetroReview
                         title="Continue Doing"
-                        items={boardData?.comments?.continueDoing || []}
-                        onAdd={(text) => addComment('continueDoing', text)}
+                        data={retroData?.reviews?.continueDoing || []}
+                        addReview={(text) => addReview('continueDoing', text)}
+                        updateReview={(text, index) => updateReview('continueDoing', text, index)}
                     />
-                </Grid>
-                <Grid item xs={12} md={3}>
-                    <RetroColumn
+
+                    <RetroReview
                         title="Appreciation"
-                        items={boardData?.comments?.appreciation || []}
-                        onAdd={(text) => addComment('appreciation', text)}
+                        data={retroData?.reviews?.appreciation || []}
+                        addReview={(text) => addReview('appreciation', text)}
+                        updateReview={(text, index) => updateReview('appreciation', text, index)}
                     />
                 </Grid>
-            </Grid>
+            </Card>
         </Container>
     );
-}
-
+};
